validate VirtualBox environment & bump min version
Introduce virtualbox_env.sh to verify that `VBoxManage` is available,
that VirtualBox meets the minimum supported version, and set the
variable VBM as a shorthand for `VBoxManage`.  If `VBoxManage` is not
available, then exit.  Otherwise, if the detected version does not meet
the minimum supported version, a warning is emitted.

Also, the minimum supported version is now bumped to be >= 4.3.x.