Support VLANs for MTU setting
For the case where we wish to set an MTU on a VLAN interface, we need
to model the underlying physical interface so we can set the MTU there
first. We call a physical interface associated with a VLAN the "parent
interface".

Configure the system with an 'inet manual' interface fragment file for
the parent interface sufficient just to set the MTU. Also set MTUs
live on parent interfaces before setting on VLANs.